Avoid "Text file busy" during download/injection
Don't download/inject to the same inode, in case it is already in use.

This PR takes the conservative approach of writing to a tempfile and then
renaming over the original in a second step.  This means a later (failed)
download/inject won't corrupt any existing file.

Fixes "Text file busy" error:

```
16:57:57 info Inject binary
16:57:57 info done exec
16:57:58 info Done injecting binary
16:57:58 info done inject
16:57:58 info done injecting
16:57:58 info Inject Error: sh: 1: cannot create /usr/local/bin/devpod: Text file busy
16:57:58 info write |1: file already closed
```