fix: Support Ports and EnvFrom in kubernetes manifest
PR #547 added support for a `podManifestTemplate` which claimed to fix
Issue #390 (although not via `portsAttributes` but via the template).

I suspect `ports` in the manifest worked in that PR and were broken by
https://github.com/loft-sh/devpod/commit/4be34b237a9653630ae687453b72117908bdc854
which did some good refactoring, but made it necessary to copy the
parameters you wanted rather than basing the pod on the manifest
directly.

I would like both ports: and envFrom: to work, and this commit fixes
both of those to come from the template. If they are unspecified in
the manifest they will be unset in the created pod too.