fix: fix recreate logic if workspace is local folder
If a workspace is created from an actual local folder (as opposed to downloading
a git repo) _and_ we're using a local provider, then `--recreate` should not try
to download that folder again.

Without this fix I was seeing an EOF issue on recreate (likely files were trying
to be copied to / from the same destination leading to file corruption). For
example, my `.git/index` file was being corrupted by this issue, leading to git
commands causing a segfault.