fix/POD 837 remove fallback cloning (#1232)
* fix(cli): only setup platform access if we're using pro provider to create
workspace

* fix(cli): remove local clone fallback

Removes the current behaviour of falling back to a local clone and
upload to VM/Pod afterwards. Over the course of the last year we've
found it is more confusing than it helps.
Instead, this PR improves the logs when cloning on the remote fails to
allow users to address the root cause better.