fix(pro): local folder upload works properly now with pro
The problem was that we correctly put the uploaded local folder under
the workspace directory /var/lib/loft/devpod/$WORKSPACE_ID/agent/.. but
when we established the SSH connection we attempted to create the actual
local workspace folder on the runner side.
Because we usually only have write access to /var/lib/loft remotely this
causes permission issues and `devpod ssh` to fail.